Create folder "firebase"
